stt-server/requirements.txt
vosk
flask
flask-socketio
sounddevice
